+++ src/flow/factory/mod.rs
@@ -3,7 +3,13 @@
 mod builtin_nodes;
 mod coding_nodes;
 mod node_factory;
+mod register_builtin;
+mod register_harness;
+mod registry;
 
 pub use builtin_nodes::*;
 pub use coding_nodes::*;
 pub use node_factory::{DefaultNodeFactory, NodeFactory};
+pub use register_builtin::register_builtin_nodes;
+pub use register_harness::register_harness_nodes;
+pub use registry::NodeRegistry;